Capture workspaces enabled and use it on selector conditionally
