+++ packages/dui3/store/accounts.ts
@@ -17,6 +17,7 @@ import { setContext } from '@apollo/client/link/context'
 import { useHostAppStore } from '~/store/hostApp'
 import type { ToastNotification } from '@speckle/ui-components'
 import { ToastNotificationType } from '@speckle/ui-components'
+import { serverInfoQuery } from '~/lib/graphql/mutationsAndQueries'
 
 export type DUIAccount = {
   /** account info coming from the host app */
@@ -25,6 +26,7 @@ export type DUIAccount = {
   client: ApolloClient<unknown>
   /** whether an intial serverinfo query succeeded. */
   isValid: boolean
+  workspacesEnabled: boolean
 }
 
 const accountTestQuery = gql`
@@ -154,11 +156,15 @@ export const useAccountStore = defineStore('accountStore', () => {
         }
       })
 
+      // get workspace enabled flag and store it in account
+      const res = await client.query({ query: serverInfoQuery })
+
       apolloClients[acc.id] = client
       newAccs.push({
         accountInfo: acc,
         client,
-        isValid: true
+        isValid: true,
+        workspacesEnabled: !!res.data.serverInfo.workspaces.workspacesEnabled
       })
     }
